Move Errors beans to common module
