+++ shopping-cart/app/shoppingcart/checkout/confirmation/CheckoutConfirmationController.java
@@ -16,7 +16,7 @@ import play.mvc.Http;
 import play.mvc.Result;
 import shoppingcart.common.CartController;
 import shoppingcart.CartSessionUtils;
-import shoppingcart.ErrorsBean;
+import common.errors.ErrorsBean;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
